Revert "-"
This reverts commit d9ac65f4b3f5f69de42dc740aa3748194b6cf84e.

+++ update.py
@@ -12,38 +12,44 @@ def get_latest_release_info():
     return response.json() if response.ok else None
 
 def download_and_install_update(url, version):
-    local_filename = 'Mixpdf.exe'
-    temp_filename = 'Mixpdf_new.exe'  
+    temp_filename = 'Mixpdf_new.exe'
+    final_filename = 'Mixpdf.exe'
+
     urlretrieve(url, temp_filename)
 
     current_app_path = os.path.abspath(sys.argv[0])
-    directory = os.path.dirname(current_app_path)
-    final_path = os.path.join(directory, local_filename)
+    directory = os.path.dirname(current_app_path)  
+    final_path = os.path.join(directory, final_filename)  # Caminho final do arquivo
 
-    if os.path.exists(final_path):
-        os.remove(final_path) 
-    os.rename(temp_filename, final_path)  
+    if current_app_path.endswith('.exe'):
+        if os.path.exists(final_path):
+            os.remove(final_path)  
+        os.rename(temp_filename, final_path)  
 
-    # Atualizando a versão registrada
-    ConfigManager.save_version(version)
+        # Atualizando a versão registrada
+        ConfigManager.save_version(version)
+
+        # Reiniciando a aplicação
+        subprocess.Popen([final_path], cwd=directory)
+        sys.exit()
+    else:
+        # Atualização para ambiente de desenvolvimento, não reiniciar
+        print("Atualização baixada. Reinicie manualmente o aplicativo.")
 
 def check_for_updates():
     current_version = ConfigManager.load_version()
-    latest_release_info = get_latest_release_info()
+    latest_release = get_latest_release_info()
 
-    if latest_release_info:
-        latest_version = latest_release_info['tag_name'].replace('V', '').replace('v', '')  # Remove 'V' ou 'v' da tag
-        assets = latest_release_info['assets']
+    if latest_release:
+        latest_version = latest_release['tag_name'].replace('V', '')
+        assets = latest_release['assets']
         browser_download_url = None
 
         for asset in assets:
-            if 'mixpdf.exe' in asset['name'].lower():  # Procura por 'mixpdf.exe' no nome do asset, ignorando maiúsculas
+            if 'Mixpdf' in asset['name']:
                 browser_download_url = asset['browser_download_url']
-                break  # Sai do loop assim que encontrar o primeiro match
 
         if latest_version != current_version and browser_download_url:
-            download_and_install_update(browser_download_url, latest_version)
-        else:
-            print("Você já está na versão mais recente ou o arquivo não foi encontrado.")
+            return latest_version, browser_download_url
 
-check_for_updates()
+    return None, None
